refactor(tests): cherry pick changes for AbstractServerlessControl.js
This was originally introduced for AWS Fargate support but is now merged
early into master before merging the fargate feature branch.
