+++ packages/serverless/test/config.js
@@ -1,24 +1,5 @@
 'use strict';
 
-exports.backendProtocol = 'https';
-exports.backendHost = 'localhost';
-exports.backendPort = 8443;
-exports.backendPath = '/serverless';
-
-exports.backendBaseUrl = `${exports.backendProtocol}://${exports.backendHost}:${exports.backendPort}${
-  exports.backendPath
-}`;
-
-exports.instanaAgentKey = 'dummy-key';
-
-exports.downstreamDummyProtocol = 'http';
-exports.downstreamDummyHost = 'localhost';
-exports.downstreamDummyPort = 3456;
-
-exports.downstreamDummyUrl = `${exports.downstreamDummyProtocol}://${exports.downstreamDummyHost}:${
-  exports.downstreamDummyPort
-}/`;
-
 exports.getAppStdio = function getAppStdio() {
   if (process.env.WITH_STDOUT || process.env.CI) {
     return [process.stdin, process.stdout, process.stderr, 'ipc'];
